Limit updates to once per day
When running frequently it is possible to hit the github releases API rate limit.

This puts the current version into the toolcache and limits updates to
once per day. Allowing people to stay on the latest version within a
day, and still run frequently without hitting the rate limit.

For: https://github.com/planetscale/setup-pscale-action/issues/12